+++ debug/flexvue/core/asynctask.class.js
@@ -19,10 +19,26 @@ export default class AsyncTask {
     return __awaiter(this, arguments, void 0, function* (method, url, params = {}, ...options) {
       const _method = method.toUpperCase();
       let redirect_url = url;
-      let headers = options.find((opt) => typeof opt === 'object' && !Array.isArray(opt));
-      let otherOptions = options.filter((opt) => opt !== headers);
-      otherOptions = otherOptions.filter((opt) => opt !== null);
-      headers = headers || (_method == 'GET' ? { 'Content-Type': 'text/plain' } : { 'Content-Type': 'application/json' });
+      let headers = _method == 'GET' ? { 'Content-Type': 'text/plain' } : { 'Content-Type': 'application/json' };
+      let _options = method == 'GET' ? params !== null ? [params] : options : options;
+      Log.d('_options', _options);
+      let otherOptions = {};
+      if (_options !== null) {
+        _options.forEach((opt) => {
+          if (opt !== null) {
+            Object.entries(opt).forEach(([key, value]) => {
+              let _startString = key.substring(0, 1);
+              if (/^[A-Z]+$/.test(_startString)) {
+                headers[key] = value;
+              } else
+              {
+                otherOptions[key] = value;
+              }
+            });
+          }
+        });
+      }
+      Log.d('headers', headers);
       let requestOptions = {
         method: _method,
         cache: 'no-cache',
@@ -33,13 +49,10 @@ export default class AsyncTask {
       if (_method != 'GET') {
         requestOptions.body = JSON.stringify(params);
       }
-      otherOptions.forEach((opt) => {
-        if (opt !== null) {
-          Object.entries(opt).forEach(([key, value]) => {
-            requestOptions[key] = value;
-          });
-        }
+      Object.entries(otherOptions).forEach(([key, value]) => {
+        requestOptions[key] = value;
       });
+      Log.d('requestOptions', requestOptions);
       const response = yield fetch(redirect_url, requestOptions);
       const contentType = response.headers.get('content-type');
       if (!response.ok) {