+++ src/example/asynctask/asynctask.ts
@@ -22,13 +22,12 @@ const onReady = () : void =>
         post_params.append(key, value as string);
     });
 
-    // AsyncTask, 옵션 | 헤더
+    // POST, 옵션+헤더
     new AsyncTask().execute('POST','http://127.0.0.1:5500/test/list',post_params,{
             mode : 'cors', // no-cors, cors, *same-origin
-            //redirect : 'follow'
-        },{
-            'Content-Type' : 'application/json',
-            // 'Content-Type': 'application/x-www-form-urlencoded',
+            redirect : 'follow',
+            // 'Content-Type' : 'application/json',
+            'Content-Type': 'application/x-www-form-urlencoded',
             'Authorization-Access-Token':'Y29tZmFuY3l1cHN'
         }
     )
@@ -42,9 +41,12 @@ const onReady = () : void =>
     });
 
 
-    // AsyncTask , 헤더
-    new AsyncTask().execute('GET','http://127.0.0.1:5500/ddd/list',{},{
-            'Content-Type' : 'application/json'
+    // POST , 옵션, 헤더
+    new AsyncTask().execute('POST','http://127.0.0.1:5500/ddd/list',{a:1,b:2},
+        {'mode':'origin',},
+        {
+            'Content-Type' : 'application/json',
+            'Access-Token' : 'ddddd'
         }
     )
     .then( data =>{
@@ -55,8 +57,8 @@ const onReady = () : void =>
         Log.e(e);
     });
 
-    // AsyncTask , 헤더+옵션
-    new AsyncTask().execute('GET','http://127.0.0.1:5500/ddd/view/1',{},{
+    // GET , 헤더+옵션
+    new AsyncTask().execute('GET','http://127.0.0.1:5500/ddd/view/1',{
             mode : 'cors',
             'Content-Type' : 'application/json'
         }
@@ -70,7 +72,7 @@ const onReady = () : void =>
     });
 
 
-    // import js
+    // // import js
     new AsyncTask().doImport('../../v1/js/values/arrays.js')
     .then( data =>{
         const resp = data;